# feat: replace sha1 and tiny-keccak with sha-1 and sha3