--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,21 +1,14 @@
 [package]
 name = "multihash"
-description = "Implementation of the multihash format"
-repository = "https://github.com/multiformats/rust-multihash"
-
-keywords = ["multihash", "ipfs"]
-
 version = "0.9.4"
-
 authors = ["dignifiedquire <dignifiedquire@gmail.com>"]
-
-license = "MIT"
-
-readme = "README.md"
-
-documentation = "https://docs.rs/multihash/"
-
 edition = "2018"
+license = "MIT"
+readme = "README.md"
+description = "Implementation of the multihash format"
+documentation = "https://docs.rs/multihash/"
+repository = "https://github.com/multiformats/rust-multihash"
+keywords = ["multihash", "ipfs"]
 
 [dependencies]
 blake2b_simd = { version = "0.5.9", default-features = false }
@@ -26,4 +19,10 @@ tiny-keccak = "1.4"
 unsigned-varint = "0.3"
 
 [dev-dependencies]
+criterion = "0.3"
+rand = "0.7"
 quickcheck = "0.9.2"
+
+[[bench]]
+name = "multihash"
+harness = false