--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,9 +13,10 @@ keywords = ["multihash", "ipfs"]
 [dependencies]
 blake2b_simd = { version = "0.5.9", default-features = false }
 blake2s_simd = { version = "0.5.9", default-features = false }
-sha1 = "0.5"
-sha2 = { version = "0.7", default-features = false }
-tiny-keccak = "1.4"
+digest = { version = "0.8", default-features = false }
+sha-1 = { version = "0.8", default-features = false }
+sha2 = { version = "0.8", default-features = false }
+sha3 = { version = "0.8", default-features = false }
 unsigned-varint = "0.3"
 
 [dev-dependencies]