--- a/src/lib.rs
+++ b/src/lib.rs
@@ -6,41 +6,42 @@
 //! A `MultihashRef` is the same as a `Multihash`, except that it doesn't own its data.
 //!
 
+#![deny(missing_docs)]
+
 mod errors;
 mod hashes;
 mod storage;
 
+use std::cmp;
 use std::convert::TryFrom;
-use std::fmt::Debug;
+use std::fmt::{self, Debug};
 use std::hash;
 
 use blake2b_simd::{blake2b, Params as Blake2bVariable};
 use blake2s_simd::{blake2s, Params as Blake2sVariable};
-use sha2::Digest;
-use tiny_keccak::Keccak;
+use digest::Digest;
 use unsigned_varint::{decode, encode};
 
-pub use errors::{DecodeError, DecodeOwnedError, EncodeError};
-pub use hashes::Hash;
-use std::{cmp, fmt};
-use storage::Storage;
+pub use self::errors::{DecodeError, DecodeOwnedError, EncodeError};
+pub use self::hashes::Hash;
+use self::storage::Storage;
 
-// Helper macro for encoding input into output using sha1, sha2, tiny_keccak, or blake2
+// Helper macro for encoding input into output using sha1, sha2, sha3, or blake2
 macro_rules! encode {
-    (sha1, Sha1, $input:expr, $output:expr) => {{
-        let mut hasher = sha1::Sha1::new();
-        hasher.update($input);
-        $output.copy_from_slice(&hasher.digest().bytes());
+    (sha1, $algorithm:ident, $input:expr, $output:expr) => {{
+        let mut hasher = sha1::$algorithm::default();
+        hasher.input($input);
+        $output.copy_from_slice(&hasher.result().as_ref());
     }};
     (sha2, $algorithm:ident, $input:expr, $output:expr) => {{
         let mut hasher = sha2::$algorithm::default();
         hasher.input($input);
         $output.copy_from_slice(hasher.result().as_ref());
     }};
-    (tiny, $constructor:ident, $input:expr, $output:expr) => {{
-        let mut kec = Keccak::$constructor();
-        kec.update($input);
-        kec.finalize($output);
+    (sha3, $algorithm:ident, $input:expr, $output:expr) => {{
+        let mut hasher = sha3::$algorithm::default();
+        hasher.input($input);
+        $output.copy_from_slice(hasher.result().as_ref());
     }};
     (blake2, $algorithm:ident, $input:expr, $output:expr) => {{
         let hash = $algorithm($input);
@@ -117,14 +118,14 @@ pub fn encode(hash: Hash, input: &[u8]) -> Result<Multihash, EncodeError> {
             SHA1 => sha1::Sha1,
             SHA2256 => sha2::Sha256,
             SHA2512 => sha2::Sha512,
-            SHA3224 => tiny::new_sha3_224,
-            SHA3256 => tiny::new_sha3_256,
-            SHA3384 => tiny::new_sha3_384,
-            SHA3512 => tiny::new_sha3_512,
-            Keccak224 => tiny::new_keccak224,
-            Keccak256 => tiny::new_keccak256,
-            Keccak384 => tiny::new_keccak384,
-            Keccak512 => tiny::new_keccak512,
+            SHA3224 => sha3::Sha3_224,
+            SHA3256 => sha3::Sha3_256,
+            SHA3384 => sha3::Sha3_384,
+            SHA3512 => sha3::Sha3_512,
+            Keccak224 => sha3::Keccak224,
+            Keccak256 => sha3::Keccak256,
+            Keccak384 => sha3::Keccak384,
+            Keccak512 => sha3::Keccak512,
             Blake2b512 => blake2::blake2b,
             Blake2b256 => blake2_256::Blake2bVariable,
             Blake2s256 => blake2::blake2s,