* Add benchmark for digest
* Replace `sha1` and `tiny-keccak` with `sha-1` and `sha3`