@koushiro 2020-02-25
# Benchmark

The results are not significantly different on my machine.

## before

```
encode/identity         time:   [132.38 ns 132.90 ns 133.41 ns]                            
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
encode/sha1             time:   [2.0473 us 2.0528 us 2.0603 us]                         
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
encode/sha2_256         time:   [4.7345 us 4.7444 us 4.7558 us]                             
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low severe
  7 (7.00%) high mild
  3 (3.00%) high severe
encode/sha2_512         time:   [3.1024 us 3.1159 us 3.1335 us]                             
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) low severe
  4 (4.00%) low mild
  4 (4.00%) high mild
  6 (6.00%) high severe
encode/sha3_224         time:   [3.9798 us 3.9823 us 3.9854 us]                             
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
encode/sha3_256         time:   [3.9744 us 3.9776 us 3.9812 us]                             
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
encode/sha3_384         time:   [5.0447 us 5.0615 us 5.0808 us]                             
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high severe
encode/keccak_224       time:   [3.9861 us 3.9904 us 3.9956 us]                               
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
encode/keccak_256       time:   [3.9549 us 3.9580 us 3.9618 us]                               
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
encode/keccak_384       time:   [4.9629 us 4.9729 us 4.9843 us]                                
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
encode/keccak_512       time:   [7.4919 us 7.4974 us 7.5035 us]                               
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
encode/blake2b_256      time:   [1.3418 us 1.3426 us 1.3434 us]                                
Found 10 outliers among 100 measurements (10.00%)
  7 (7.00%) high mild
  3 (3.00%) high severe
encode/blake2b_512      time:   [1.2973 us 1.2994 us 1.3014 us]                                
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
encode/blake2s_128      time:   [1.9659 us 1.9675 us 1.9692 us]                                
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
encode/blake2s_256      time:   [2.0079 us 2.0097 us 2.0117 us]                                
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high mild
```

## after

```
encode/identity         time:   [126.36 ns 126.48 ns 126.58 ns]                            
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) low mild
encode/sha1             time:   [2.0177 us 2.0274 us 2.0377 us]                         
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
encode/sha2_256         time:   [4.8577 us 4.8865 us 4.9173 us]                             
encode/sha2_512         time:   [3.1260 us 3.1373 us 3.1510 us]                             
encode/sha3_224         time:   [4.0874 us 4.1135 us 4.1395 us]                             
encode/sha3_256         time:   [4.0373 us 4.0502 us 4.0663 us]                             
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
encode/sha3_384         time:   [5.0627 us 5.0666 us 5.0712 us]                             
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
encode/keccak_224       time:   [4.0856 us 4.1092 us 4.1340 us]                               
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
encode/keccak_256       time:   [4.1095 us 4.1300 us 4.1487 us]                               
encode/keccak_384       time:   [5.0831 us 5.0890 us 5.0961 us]                               
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
encode/keccak_512       time:   [7.5094 us 7.5458 us 7.5827 us]                               
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
encode/blake2b_256      time:   [1.3323 us 1.3336 us 1.3351 us]                                
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe
encode/blake2b_512      time:   [1.2899 us 1.2919 us 1.2941 us]                                
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
encode/blake2s_128      time:   [2.0425 us 2.0441 us 2.0460 us]                                
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low severe
  1 (1.00%) high mild
  1 (1.00%) high severe
encode/blake2s_256      time:   [2.0341 us 2.0633 us 2.0953 us]                                
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```