@vmx 2020-02-25
Thanks for adding benchmarks. With the merge of https://github.com/multiformats/rust-multihash/pull/45 the codebase changed quite a lot. Could you please rebase the benchmarks?

Why switching crates if there is not really a perf difference?