@koushiro 2020-02-25
> Thanks for adding benchmarks. With the merge of #45 the codebase changed quite a lot. Could you please rebase the benchmarks?
> 
> Why switching crates if there is not really a perf difference?

Because `sha-1`, `sha2` and `sha3` are a series of libraries that have consistent API.